trinetra-backend/requirements.txt
# Trinetra Backend — Python Dependencies
# ═══ Core Framework ═══
fastapi==0.115.12
uvicorn[standard]==0.34.3

# ═══ Firebase ═══
firebase-admin==6.7.0

# ═══ HTTP Client (for scrapers & Firebase REST API) ═══
httpx==0.28.1

# ═══ Web Scraping ═══
beautifulsoup4==4.13.4
lxml==5.4.0

# ═══ Data Validation ═══
pydantic==2.11.4
pydantic-settings==2.9.1

# ═══ Environment Variables ═══
python-dotenv==1.1.0
